handlers/requirements.txt
requests
flask-lambda-python36
pytest
wheel
aws-sam-cli
awscli
